Next batch operation registry moves
